+++ CHANGELOG.md
@@ -13,8 +13,11 @@ and this project adheres to
   fields you want to change ([#1309]).
 - @cosmjs/tendermint-rpc: Add missing exports `HttpBatchClient`,
   `HttpBatchClientOptions`, `RpcClient` ([#1311]).
+- @cosmjs/tendermint-rpc: Send batch immediately when full in `HttpBatchClient`
+  ([#1310]).
 
 [#1309]: https://github.com/cosmos/cosmjs/issues/1309
+[#1310]: https://github.com/cosmos/cosmjs/issues/1310
 [#1311]: https://github.com/cosmos/cosmjs/issues/1311
 
 ### Fixed